Adding support to sessions in the thematic tool

+++ main/attendance/attendance_calendar.php
@@ -14,15 +14,18 @@ $param_gradebook = '';
 if (isset($_SESSION['gradebook'])) {
 	$param_gradebook = '&gradebook='.Security::remove_XSS($_SESSION['gradebook']);
 }
-echo '<div class="actions" style="margin-bottom:30px">';
-echo '<a href="index.php?'.api_get_cidreq().'&action=attendance_sheet_list&attendance_id='.$attendance_id.$param_gradebook.'">'.Display::return_icon('attendance_list.png',get_lang('AttendanceSheet'),'','32').'</a>';
-if ($action == 'calendar_add') {
-	echo '<a href="index.php?'.api_get_cidreq().'&action=calendar_list&attendance_id='.$attendance_id.$param_gradebook.'">'.Display::return_icon('attendance_calendar.png',get_lang('AttendanceCalendar'),'','32').'</a>';
-} else {
-	echo '<a href="index.php?'.api_get_cidreq().'&action=calendar_add&attendance_id='.$attendance_id.$param_gradebook.'">'.Display::return_icon('add_calendar_event.png',get_lang('AddDateAndTime'),'','32').'</a>';
-	echo '<a onclick="javascript:if(!confirm(\''.get_lang('AreYouSureToDeleteAllDates').'\')) return false;" href="index.php?'.api_get_cidreq().'&action=calendar_all_delete&attendance_id='.$attendance_id.$param_gradebook.'">'.Display::return_icon('clean_calendar.png',get_lang('CleanCalendar'),'','32').'</a>';
+if (!$is_locked_attendance || api_is_platform_admin()) {
+	echo '<div class="actions" style="margin-bottom:30px">';
+	echo '<a href="index.php?'.api_get_cidreq().'&action=attendance_sheet_list&attendance_id='.$attendance_id.$param_gradebook.'">'.Display::return_icon('attendance_list.png',get_lang('AttendanceSheet'),'','32').'</a>';
+	if ($action == 'calendar_add') {
+		echo '<a href="index.php?'.api_get_cidreq().'&action=calendar_list&attendance_id='.$attendance_id.$param_gradebook.'">'.Display::return_icon('attendance_calendar.png',get_lang('AttendanceCalendar'),'','32').'</a>';
+	} else {
+		echo '<a href="index.php?'.api_get_cidreq().'&action=calendar_add&attendance_id='.$attendance_id.$param_gradebook.'">'.Display::return_icon('add_calendar_event.png',get_lang('AddDateAndTime'),'','32').'</a>';
+		echo '<a onclick="javascript:if(!confirm(\''.get_lang('AreYouSureToDeleteAllDates').'\')) return false;" href="index.php?'.api_get_cidreq().'&action=calendar_all_delete&attendance_id='.$attendance_id.$param_gradebook.'">'.Display::return_icon('clean_calendar.png',get_lang('CleanCalendar'),'','32').'</a>';
+	}
+	echo '</div>';
 }
-echo '</div>';
+
 
 $message_information = get_lang('AttendanceCalendarDescription');
 
@@ -70,8 +73,8 @@ if (isset($action) && $action == 'calendar_add') {
                 $defaults['end_date_time'] = date('Y-m-d 12:00:00');
                 $form->addElement('html', '</td></tr>');
                 $form->addElement('html', '</table>');
-
-
+                
+                
                 $form->addElement('html', '</div>');
                 $form->addElement('html', '</div>');
                 $form->addElement('html','<br /><br />');
@@ -85,32 +88,33 @@ if (isset($action) && $action == 'calendar_add') {
 	// calendar list
 	echo '<div class="attendance-calendar-list">';
 	echo '<h4>'.get_lang('CalendarList').'</h4>';
-	if (!empty($attendance_calendar)) {
-		foreach ($attendance_calendar as $calendar) {
-			echo '<div class="attendance-calendar-row">';
+	if (!empty($attendance_calendar)) {	
+		foreach ($attendance_calendar as $calendar) {	
+			echo '<div class="attendance-calendar-row">';				
 				if ((isset($action) && $action == 'calendar_edit') && (isset($calendar_id) && $calendar_id == $calendar['id'])) {
-					// calendar edit form
-					echo '<div class="attendance-calendar-edit">';
-						$form = new FormValidator('attendance_calendar_edit','POST','index.php?action=calendar_edit&attendance_id='.$attendance_id.'&calendar_id='.$calendar_id.'&'.api_get_cidreq().$param_gradebook,'');
+					// calendar edit form				
+					echo '<div class="attendance-calendar-edit">';				
+						$form = new FormValidator('attendance_calendar_edit','POST','index.php?action=calendar_edit&attendance_id='.$attendance_id.'&calendar_id='.$calendar_id.'&'.api_get_cidreq().$param_gradebook,'');					
 						$form->addElement('datepicker', 'date_time', '', array('form_name'=>'attendance_calendar_edit'), 5);
 						$defaults['date_time'] = $calendar['date_time'];
 						$form->addElement('style_submit_button', null, get_lang('Save'), 'class="save"');
 						$form->addElement('style_submit_button', 'cancel', get_lang('Cancel'), 'class="cancel"');
 						$form->setDefaults($defaults);
 						$form->display();
-					echo '</div>';
+					echo '</div>';						
 				} else {
 					echo Display::return_icon('lp_calendar_event.png',get_lang('DateTime')).' '.$calendar['date_time'].'&nbsp;';
-					echo '<span style="margin-left:20px;">';
-					echo '<a href="index.php?'.api_get_cidreq().'&action=calendar_edit&calendar_id='.intval($calendar['id']).'&attendance_id='.$attendance_id.$param_gradebook.'">'.Display::return_icon('edit.gif', get_lang('Edit'), array('style'=>'vertical-align:middle')).'</a>&nbsp;';
-					echo '<a onclick="javascript:if(!confirm(\''.get_lang('AreYouSureToDelete').'\')) return false;" href="index.php?'.api_get_cidreq().$param_gradebook.'&action=calendar_delete&calendar_id='.intval($calendar['id']).'&attendance_id='.$attendance_id.'">'.Display::return_icon('delete.gif', get_lang('Delete'), array('style'=>'vertical-align:middle')).'</a>';
-					echo '</span>';
-				}
+					if (!$is_locked_attendance || api_is_platform_admin()) {
+                        echo '<span style="margin-left:20px;">';
+                        echo '<a href="index.php?'.api_get_cidreq().'&action=calendar_edit&calendar_id='.intval($calendar['id']).'&attendance_id='.$attendance_id.$param_gradebook.'">'.Display::return_icon('edit.gif', get_lang('Edit'), array('style'=>'vertical-align:middle')).'</a>&nbsp;';
+                        echo '<a onclick="javascript:if(!confirm(\''.get_lang('AreYouSureToDelete').'\')) return false;" href="index.php?'.api_get_cidreq().$param_gradebook.'&action=calendar_delete&calendar_id='.intval($calendar['id']).'&attendance_id='.$attendance_id.'">'.Display::return_icon('delete.gif', get_lang('Delete'), array('style'=>'vertical-align:middle')).'</a>';
+                        echo '</span>';
+                    }
+				}	
 			echo '</div>';
 		}
 	} else {
 		echo '<div style="padding:5px;">'.get_lang('ThereAreNoRegisteredDatetimeYet').'</div>';
 	}
-	echo '</div>';
+	echo '</div>';			
 }
-?>
